+++ rpc-framework/rpc-common/src/main/java/space/ruiwang/utils/serializer/JsonBinarySerializer.java
@@ -1,4 +1,4 @@
-package space.ruiwang.utils;
+package space.ruiwang.utils.serializer;
 
 import java.nio.charset.StandardCharsets;
 
@@ -10,12 +10,12 @@ import cn.hutool.json.JSONUtil;
  */
 public class JsonBinarySerializer {
     // JSON序列化
-    public static byte[] serializeToJsonBytes(Object object) {
+    public static byte[] serialize(Object object) {
         return JSONUtil.toJsonStr(object).getBytes(StandardCharsets.UTF_8);
     }
 
     // JSON反序列化
-    public static <T> T deserializeFromJsonBytes(byte[] jsonBytes, Class<T> valueType) {
+    public static <T> T deserialize(byte[] jsonBytes, Class<T> valueType) {
         String jsonStr = new String(jsonBytes, StandardCharsets.UTF_8);
         return JSONUtil.toBean(jsonStr, valueType);
     }